fix(ci): split internal/watch/watch.go under the 800-line hygiene gate, fix ST1005 lint
watch.go grew to 866 lines from #508's dedup/cap/TTL logic, tripping the
workflow-hygiene file-size gate. Split along the existing type/persistence
seam: watch.go keeps the Watch/PricePoint types, validation, and display
helpers (Sparkline, TrendArrow); store.go gets the Store type and all
persistence/pruning methods plus the scaling-guard consts they use. Pure
move, no logic changes -- build/vet/staticcheck/full test suite all green
before and after.

Also fixes the unrelated ST1005 staticcheck failure in watch_daemon.go
(error string ending in punctuation).

The OSV grpc CVE that was also failing on this branch is already fixed on
main (1.82.1); rebasing to pick it up.
